+++ src/const.js
@@ -1,0 +1,5 @@
+const FILM_CARD_COUNT_PER_STEP = 5;
+const FILM_CARD_MOCK_COUNT = 20;
+const TOP_COMMENTED_FILM_CARD_COUNT = 2;
+
+export {FILM_CARD_COUNT_PER_STEP, FILM_CARD_MOCK_COUNT, TOP_COMMENTED_FILM_CARD_COUNT};
