+++ src/main.js
@@ -1,5 +1,5 @@
 import { createUserStatusTemplate } from './view/user-status';
-import { createSiteMenuTemplate, createStatsTemplate } from './view/site-menu-and-stats';
+import { createSiteMenuTemplate } from './view/site-menu';
 import { createSortingTemplate } from './view/sorting';
 import { createAllFilmsListTemplate } from './view/all-films-list';
 import { createShowMoreButtonTemplate } from './view/show-more-button';
@@ -7,15 +7,15 @@ import { createTopFilmsTemplate } from './view/top-rated-films';
 import { createMostCommentedFilmsTemplate } from './view/most-commented-films';
 import { createFilmCardTemplate } from './view/film-card';
 import { createFilmPopupTemplate } from './view/film-pop-up';
+import { createCommentTemplate } from './view/comments-in-popup';
+import { createPopupGenreTemplate } from './view/film-genre';
 import { generateFilm } from './mock/film';
-import { generateFilter } from './mock/filter-and-stats';
-
-const FILM_CARD_COUNT_PER_STEP = 5;
-const FILM_CARD_MOCK_COUNT = 20;
-const TOP_COMMENTED_FILM_CARD_COUNT = 2;
+import { generateFilter } from './mock/filter';
+import { createStatsTemplate } from './view/stats.js';
+import { FILM_CARD_COUNT_PER_STEP, FILM_CARD_MOCK_COUNT, TOP_COMMENTED_FILM_CARD_COUNT } from './const';
 
 const films = new Array(FILM_CARD_MOCK_COUNT).fill().map(generateFilm);
-const filtersAndStats = generateFilter(films);
+const filters = generateFilter(films);
 
 const render = (containter, template, place) => {
   containter.insertAdjacentHTML(place, template);
@@ -25,18 +25,23 @@ const userLogoElement = document.querySelector('.header__logo');
 const siteMainElement = document.querySelector('.main');
 
 render(userLogoElement, createUserStatusTemplate(), 'beforeend');
-render(siteMainElement, createSiteMenuTemplate(filtersAndStats), 'beforeend');
+render(siteMainElement, createSiteMenuTemplate(filters), 'beforeend');
 render(siteMainElement, createSortingTemplate(), 'beforeend');
 
 const filmsElement = document.querySelector('.films');
 
 render(filmsElement, createAllFilmsListTemplate(), 'beforeend');
 const allFilmsContainerElement = document.querySelector('.films-list__container');
-for (let i = 0; i < Math.min(films.length, FILM_CARD_COUNT_PER_STEP); i++) {
-  render(allFilmsContainerElement, createFilmCardTemplate(films[i]), 'beforeend');
-}
+films.slice(0,FILM_CARD_COUNT_PER_STEP).forEach((film) => render(allFilmsContainerElement, createFilmCardTemplate(film), 'beforeend'));
 
 render(filmsElement, createFilmPopupTemplate(films[0]), 'beforeend');
+
+const genreInPopupContainterElement = document.querySelector('.film-details__table');
+render(genreInPopupContainterElement, createPopupGenreTemplate(films[0]), 'beforeend');
+
+
+const commentsInPopupContainterElement = document.querySelector('.film-details__bottom-container');
+render(commentsInPopupContainterElement, createCommentTemplate(films[0]), 'beforeend');
 
 if (films.length > FILM_CARD_COUNT_PER_STEP) {
   let renderedFilmsCount = FILM_CARD_COUNT_PER_STEP;
@@ -61,16 +66,12 @@ if (films.length > FILM_CARD_COUNT_PER_STEP) {
 
 render(filmsElement, createTopFilmsTemplate(), 'beforeend');
 const topFilmsContainerElement = document.querySelectorAll('.films-list__container')[1];
-for (let i = 0; i < TOP_COMMENTED_FILM_CARD_COUNT; i++) {
-  render(topFilmsContainerElement, createFilmCardTemplate(films[i]), 'beforeend');
-}
+films.slice(0,TOP_COMMENTED_FILM_CARD_COUNT).forEach((film) => render(topFilmsContainerElement, createFilmCardTemplate(film), 'beforeend'));
 
 render(filmsElement, createMostCommentedFilmsTemplate(), 'beforeend');
 const mostCommentedFilmsContainerElement = document.querySelectorAll('.films-list__container')[2];
-for (let i = 0; i < TOP_COMMENTED_FILM_CARD_COUNT; i++) {
-  render(mostCommentedFilmsContainerElement, createFilmCardTemplate(films[i]), 'beforeend');
-}
+films.slice(0,TOP_COMMENTED_FILM_CARD_COUNT).forEach((film) => render(mostCommentedFilmsContainerElement, createFilmCardTemplate(film), 'beforeend'));
 
 const footerStaticsElement = document.querySelector('.footer__statistics');
 
-render(footerStaticsElement, createStatsTemplate(filtersAndStats[0]), 'beforeend');
+render(footerStaticsElement, createStatsTemplate(films), 'beforeend');
