+++ src/view/film-pop-up.js
@@ -1,35 +1,3 @@
-import { generateComment } from '../mock/comments';
-
-const createPopupGenreTemplate = (genre) => {
-  const popupGenreTemplate = [];
-  if (genre.length !== 0) {
-    genre.forEach((element) => popupGenreTemplate.push(`<span class="film-details__genre">${element}</span>`));
-  } else {popupGenreTemplate.push('<span class="film-details__genre"></span>');}
-
-  return popupGenreTemplate.join('');
-};
-
-const createCommentTemplate = (commentsArray) => {
-  const popupCommentsTemplate = [];
-  if(commentsArray.length !==0) {
-    commentsArray.forEach((element)=> popupCommentsTemplate.push(`<li class="film-details__comment">
-          <span class="film-details__comment-emoji">
-            <img src="./images/emoji/${element.emoji}.png" width="55" height="55" alt="emoji-${element.emoji}">
-          </span>
-          <div>
-            <p class="film-details__comment-text">${element.text}</p>
-            <p class="film-details__comment-info">
-              <span class="film-details__comment-author">${element.author}</span>
-              <span class="film-details__comment-day">${element.date}</span>
-              <button class="film-details__comment-delete">Delete</button>
-            </p>
-          </div>
-        </li>`));
-  } else {popupCommentsTemplate.push('');}
-
-  return popupCommentsTemplate.join('');
-};
-
 export const createFilmPopupTemplate = (film = {}) => {
   const {
     title = '',
@@ -43,29 +11,22 @@ export const createFilmPopupTemplate = (film = {}) => {
     releaseDate ='',
     duration = '',
     country = [],
-    genre = [],
-    comments = [],
     isAddedToWatchlist = false,
     isAlreadyWatched = false,
     isFavorite = false,
   } = film;
 
-  const commentsArray = new Array(comments.length).fill().map(generateComment);
-
   const addedToWatchlistClassName = isAddedToWatchlist
-    ? 'film-details__control-button--active film-details__control-button--watchlist'
-    : 'film-details__control-button--watchlist';
+    ? 'film-details__control-button--active '
+    : '';
 
   const alreadyWatchedClassName = isAlreadyWatched
-    ? 'film-details__control-button--active film-details__control-button--watched'
-    : 'film-details__control-button--watched';
+    ? 'film-details__control-button--active '
+    : '';
 
   const favoriteClassName = isFavorite
-    ? 'film-details__control-button--active film-details__control-button--favorite'
-    : 'film-details__control-button--favorite';
-
-  const genreTemplate = createPopupGenreTemplate(genre);
-  const commentsTemplate = createCommentTemplate(commentsArray);
+    ? 'film-details__control-button--active '
+    : '';
 
   return `<section class="film-details">
   <form class="film-details__inner" action="" method="get">
@@ -117,11 +78,6 @@ export const createFilmPopupTemplate = (film = {}) => {
               <td class="film-details__term">Country</td>
               <td class="film-details__cell">${country}</td>
             </tr>
-            <tr class="film-details__row">
-              <td class="film-details__term">${genre.length !== 1 ? 'Genres' : 'Genre'}</td>
-              <td class="film-details__cell">
-              ${genreTemplate}
-            </tr>
           </table>
 
           <p class="film-details__film-description">
@@ -131,50 +87,13 @@ export const createFilmPopupTemplate = (film = {}) => {
       </div>
 
       <section class="film-details__controls">
-        <button type="button" class="film-details__control-button ${addedToWatchlistClassName}" id="watchlist" name="watchlist">Add to watchlist</button>
-        <button type="button" class="film-details__control-button ${alreadyWatchedClassName}" id="watched" name="watched">Already watched</button>
-        <button type="button" class="film-details__control-button ${favoriteClassName}" id="favorite" name="favorite">Add to favorites</button>
+        <button type="button" class="film-details__control-button ${addedToWatchlistClassName} film-details__control-button--watchlist" id="watchlist" name="watchlist">Add to watchlist</button>
+        <button type="button" class="film-details__control-button ${alreadyWatchedClassName} film-details__control-button--watched" id="watched" name="watched">Already watched</button>
+        <button type="button" class="film-details__control-button ${favoriteClassName} film-details__control-button--favorite" id="favorite" name="favorite">Add to favorites</button>
       </section>
     </div>
 
     <div class="film-details__bottom-container">
-      <section class="film-details__comments-wrap">
-        <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${comments.length}</span></h3>
-
-        <ul class="film-details__comments-list">
-          ${commentsTemplate}
-        </ul>
-
-        <div class="film-details__new-comment">
-          <div class="film-details__add-emoji-label"></div>
-
-          <label class="film-details__comment-label">
-            <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-          </label>
-
-          <div class="film-details__emoji-list">
-            <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
-            <label class="film-details__emoji-label" for="emoji-smile">
-              <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
-            </label>
-
-            <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
-            <label class="film-details__emoji-label" for="emoji-sleeping">
-              <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
-            </label>
-
-            <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
-            <label class="film-details__emoji-label" for="emoji-puke">
-              <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
-            </label>
-
-            <input class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
-            <label class="film-details__emoji-label" for="emoji-angry">
-              <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
-            </label>
-          </div>
-        </div>
-      </section>
     </div>
   </form>
 </section>`;
