+++ src/view/stats.js
@@ -1,0 +1,3 @@
+export const createStatsTemplate = (films = {}) => (
+  `<p>${films.length} movies inside</p>`
+);
